+++ x/gov/client/cli/query.go
@@ -364,8 +364,9 @@ $ %s query gov deposit 1 cosmos1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk
 			}
 
 			// check to see if the proposal is in the store
-			_, err = queryClient.Proposal(
-				context.Background(),
+			ctx := cmd.Context()
+			proposalRes, err := queryClient.Proposal(
+				ctx,
 				&types.QueryProposalRequest{ProposalId: proposalID},
 			)
 			if err != nil {
@@ -377,25 +378,27 @@ $ %s query gov deposit 1 cosmos1skjwj5whet0lpe65qaq4rpq03hjxlwd9nf39lk
 				return err
 			}
 
-			res, err := queryClient.Deposit(
-				context.Background(),
-				&types.QueryDepositRequest{ProposalId: proposalID, Depositor: args[1]},
-			)
-			if err != nil {
-				return err
-			}
-
-			deposit := res.GetDeposit()
-			if deposit.Empty() {
+			var deposit types.Deposit
+			propStatus := proposalRes.Proposal.Status
+			if !(propStatus == types.StatusVotingPeriod || propStatus == types.StatusDepositPeriod) {
 				params := types.NewQueryDepositParams(proposalID, depositorAddr)
 				resByTxQuery, err := gcutils.QueryDepositByTxQuery(clientCtx, params)
 				if err != nil {
 					return err
 				}
 				clientCtx.JSONMarshaler.MustUnmarshalJSON(resByTxQuery, &deposit)
+				return clientCtx.PrintProto(&deposit)
 			}
 
-			return clientCtx.PrintProto(&deposit)
+			res, err := queryClient.Deposit(
+				ctx,
+				&types.QueryDepositRequest{ProposalId: proposalID, Depositor: args[1]},
+			)
+			if err != nil {
+				return err
+			}
+
+			return clientCtx.PrintProto(&res.Deposit)
 		},
 	}